[REFACTOR#1603] 인 메모리 캐시 적용 및 책임 분리
