+++ data/home/src/main/java/org/sopt/official/data/home/di/ApiModule.kt
@@ -31,7 +31,6 @@ import dagger.hilt.components.SingletonComponent
 import org.sopt.official.common.di.AppRetrofit
 import org.sopt.official.data.home.remote.api.CalendarApi
 import org.sopt.official.data.home.remote.api.HomeApi
-import org.sopt.official.data.home.remote.api.UserApi
 import retrofit2.Retrofit
 import retrofit2.create
 import javax.inject.Singleton
@@ -39,10 +38,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 internal object ApiModule {
-
-    @Provides
-    @Singleton
-    internal fun provideUserApi(@AppRetrofit(true) retrofit: Retrofit): UserApi = retrofit.create()
 
     @Provides
     @Singleton
